seccomp: change default response to ENOSYS
A common problem (e.g. with the introduction of clone3 and glibc's use of
it) with returning EPERM is that arbitrary userspace programs are not used
to understanding EPERM as "fall back to some other thing". However, they
*are* generally used to doing this when they see ENOSYS, because they are
potentially running on an older kernel which does not support whatever this
new syscall is. That is, the typical pattern for userspace is: test if
$fancy_new_syscall returns something other than ENOSYS, if so, try to use
it, if not fall back to $old_less_fancy_thing. Since EPERM != ENOSYS,
userspace ends up failing, when it could potentially still fall back to an
older implementation of the functionality.

So, let's change the default response to ENOSYS, so that userspace knows
"try something else". Since the syscall is still not run, there is no
security regression here, it just gives userspace a better hint to fall
back gracefully if it can.

This change comes out of a discussion at Linux Security Summit North
America during Tom Hromatka and Paul Moore's libseccomp talk:
https://lssna2025.sched.com/event/1zam9/handling-new-syscalls-in-seccomp-filters-tom-hromatka-oracle-corporate-paul-moore-microsoft

For the tests, I mostly just had to fix the expected errno for tests, with
the exception of some of the chroot tests, which were failing because the
seccomp policy allowed it, not because CAP_SYS_CHROOT was dropped. So that
bit of fixing makes TestUserNoEffectiveCapabilitiesChroot test the right
thing. However, it's entirely possible some additional stuff will fail, as
I couldn't run a lot of stuff locally for some reason...

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>